+++ src/domain/value/space_display_id.go
@@ -1,5 +1,5 @@
 package value
 
 type SpaceDisplayId struct {
-	Value uint
+	Value uint64
 }
