+++ src/domain/value/space_id.go
@@ -1,5 +1,5 @@
 package value
 
 type SpaceId struct {
-	Value uint
+	Value uint64
 }
